+++ example/lib/lite_state_controllers/single_use_controller.dart
@@ -10,6 +10,18 @@ class SingleUseController extends LiteStateController<SingleUseController> {
     return _counter;
   }
 
+  String? get date {
+    final date = getPersistentValue('date');
+    return date?.toString() ?? 'none';
+  }
+
+  Future setDate() async {
+    setPersistentValue(
+      'date',
+      DateTime.now(),
+    );
+  }
+
   set counter(int value) {
     _counter = value;
     rebuild();